+++ backend/config/routes.php
@@ -116,6 +116,7 @@ return [
     '/api/user/group/public'                         => ['GET',    GroupController::class.'::public'],
     '/api/user/group/create'                         => ['POST',   GroupController::class.'::create'],
     '/api/user/group/{id}/rename'                    => ['PUT',    GroupController::class.'::rename'],
+    '/api/user/group/{id}/update-description'        => ['PUT',    GroupController::class.'::updateDescription'],
     '/api/user/group/{id}/set-visibility/{choice}'   => ['PUT',    GroupController::class.'::setVisibility'],
     '/api/user/group/{id}/set-auto-accept/{choice}'  => ['PUT',    GroupController::class.'::setAutoAccept'],
     '/api/user/group/{id}/delete'                    => ['DELETE', GroupController::class.'::delete'],