+++ backend/tests/Functional/Controller/App/ApplicationControllerTest.php
@@ -49,7 +49,7 @@ class ApplicationControllerTest extends WebTestCase
         $this->assertSame([
             'id' => $app->getId(),
             'name' => 'Test App',
-            'groups' => [['id' => $group->getId(), 'name' => 'g1',
+            'groups' => [['id' => $group->getId(), 'name' => 'g1', 'description' => null,
                 'visibility' => Group::VISIBILITY_PRIVATE, 'autoAccept' => false]],
             'roles' => ['app'],
         ], $this->parseJsonBody($response));